feat: MariaDB/MySQL support  (#185)
* fix: make sure the default value of translatable attributes work in MySQL

* chore: another way of registering the LocalesType

* fix: use Arel.sql instead (in order to support Rails 7.0)

* chore(ci): update the test matrix (databases / rails versions)

* chore(ci): MySQL/MariaDB doesnt work well with Rails 7.0

+++ db/migrate/20211013210954_translate_section_content.rb
@@ -1,11 +1,13 @@
 class TranslateSectionContent < ActiveRecord::Migration[6.0]
   def change
-    remove_column :maglev_sites, :sections, :jsonb, default: []
-    remove_column :maglev_pages, :sections, :jsonb, default: []
+    remove_column :maglev_sites, :sections, :jsonb, default: [] if column_exists?(:maglev_sites, :sections)
+    remove_column :maglev_pages, :sections, :jsonb, default: [] if column_exists?(:maglev_pages, :sections)
 
     change_table :maglev_sites do |t|
       if t.respond_to? :jsonb
         t.jsonb :sections_translations, default: {}
+      elsif mysql?
+        t.json :sections_translations # MySQL doesn't support default values for json columns
       else
         t.json :sections_translations, default: {}
       end
@@ -14,6 +16,8 @@ class TranslateSectionContent < ActiveRecord::Migration[6.0]
     change_table :maglev_pages do |t|
       if t.respond_to? :jsonb
         t.jsonb :sections_translations, default: {}
+      elsif mysql?
+        t.json :sections_translations # MySQL doesn't support default values for json columns
       else
         t.json :sections_translations, default: {}
       end
